Replace SpriteText with NewSpriteText
