+++ osu.Framework/Graphics/Visualisation/LogOverlay.cs
@@ -202,7 +202,6 @@ namespace osu.Framework.Graphics.Visualisation
                     Padding = new MarginPadding { Left = target_box_width + 10 },
                     Child = new SpriteText
                     {
-                        AutoSizeAxes = Axes.Y,
                         RelativeSizeAxes = Axes.X,
                         TextSize = font_size,
                         Text = entry.Message
